requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
groq==0.9.0
pinecone-client==4.1.0
sentence-transformers==2.7.0
pymupdf==1.24.3
python-dotenv==1.0.1
